Refactor canvass map to use separate AddressMap component
