+++ src/components/sections/canvass/AllRoutesPane.scss
@@ -7,12 +7,12 @@
         width: 32em;
         z-index: 2;
     }
-}
 
-.AllRoutesPane-map {
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    top: 0;
+    .AddressMap {
+        position: absolute;
+        bottom: 0;
+        right: 0;
+        left: 0;
+        top: 0;
+    }
 }
