+++ src/components/sections/canvass/elements/AddressMap.scss
@@ -1,0 +1,11 @@
+.AddressMap {
+    position: relative;
+
+    .AddressMap-map {
+        position: absolute;
+        bottom: 0;
+        right: 0;
+        left: 0;
+        top: 0;
+    }
+}
